+++ include/lbann/objective_functions/lbann_objective_fn_categorical_cross_entropy.hpp
@@ -37,7 +37,7 @@ class categorical_cross_entropy : public objective_fn {
   ~categorical_cross_entropy();
 
   void setup(int num_neurons, int mini_batch_size);
-  void fp_set_std_matrix_view(int64_t cur_mini_batch_size);
+  void fp_set_std_matrix_view(int cur_mini_batch_size);
   double compute_categorical_cross_entropy(ElMat& predictions_v, ElMat& groundtruth_v);
   double compute_obj_fn(ElMat& predictions_v, ElMat& groundtruth_v);
   void compute_obj_fn_derivative(layer_type prev_layer_type,
